Remove check for COMPILATION_INTERRUPTED
performOptimization in OMR::Optimizer::optimize() never returns a
negative value. Nor is there any place in the code where
COMPILATION_INTERRUPTED is returned. As such, this commit removes the
check for COMPILATION_INTERRUPTED in order to maintain consistency and
remove any further use of it as a return code.

Issue: #380
Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>