Fix Mandelbrot buffer size issue
Valgrind errors were caused in Mandelbrot due to out of scope
memory writes caused by wiritng a 32-bit value `bits` in an
8-bit index at `buffer`.

The errors were caused on writes to the last three indices in
`buffer`. This commit fixes the errors by explicitly converting
`bits` to 8-bit.

The underlying problem with StoreAt needs to be fixed.

Issue: #414
Signed-off-by: Aman Kumar <amank@ca.ibm.com>