Improve methodHandleInvokeInliningGroup
These changes are part of work to clean up the handling of the
methodHandleInvokeInliningGroup and generalize it as a type of targeted
inlining. It turns off doJSR292PerfTweaks for now since this is an API
we want to eliminate in the long run but are not quite ready to delete
it completely - we just turn it off as deprecated pending final removal
at a later date. It adds a targetedInlining optimization enum which is
used to distinguish targetedInlining from regular inlining in the opt
strategies.

Task 120255
Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>