Extract Priv Arg Remat implementation for reuse
The notion of repeating stores on a cold path is a generally useful
concept for the optimizer. Currently the code used to do this is in the
liner, but the common core implementation is generally useful. This
change extract the remat implemetnation into its own file to facilitate
later reuse in other optimizations. The only funcitonal change is to
switch the logging in the extract rematerialization code to use the
common traceMsg infrastrcture rather than the inliner trace facility.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>