Store callNode ByteCodeInfo in a TR_VirtualGuard
For a given virtual guard you may want to retrieve the original
ByteCodeInfo for the guard. While the caller index and bytecode index
are stored in the inlining table, the table does not store the true
original ByteCodeInfo including the doNotProfile bit. The
VirtualGuard constructor is always given the callNode (when it
exists) so this change stashes the call's ByteCodeInfo when possible
otherwise it sets the ByteCodeInfo to be invalid with the doNotProfile
bit set.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>