Fix test compiler build errors
ResolvedMethod and Method are dependent on each other; both include
and use the other one.

The solution is to create a ResolvedMethod.cpp and move the
Method-specific stuff from ResolvedMethod.hpp to ResolvedMethod.cpp.
This way, ResolvedMethod.hpp can get away with a forward declaration
of Method instead of an include.

Add ResolvedMethod.cpp to makefiles.

Fix other missing includes.